+++ config.json
@@ -4,7 +4,7 @@
   "imgUrl": "icon.png",
   "execProgram": "node",
   "main": "app.js",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "Helm CLI wrapper",
   "category": "CONTAINERS",
   "auth": {
