+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "kaholo-plugin-kubernetes",
+  "name": "kaholo-plugin-helm",
   "description": "Kubernetes API plugin for kaholo",
   "main": "app.js",
   "scripts": {
@@ -10,10 +10,10 @@
   "license": "ISC",
   "dependencies": {
     "@kaholo/plugin-library": "^2.0.0",
-    "@kubernetes/client-node": "^0.14.0",
-    "eslint": "^8.8.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.4",
-    "js-yaml": "^4.0.0"
+    "@kubernetes/client-node": "0.14.0",
+    "eslint": "8.8.0",
+    "eslint-config-airbnb-base": "15.0.0",
+    "eslint-plugin-import": "2.25.4",
+    "js-yaml": "4.0.0"
   }
 }
